photo update
When the picture is taken in Capture.js the image is stored in a base64 encoded string then appended to formdata, then it sends a post request with the formdata in the request body.  The post method gets the string and writes it as an image.  The image is then being stored server side, in the scratch folder. But it only takes pictures in burst of 5 on my machine. Without the time stamp it will only overwrite once.